Evaluate and log changes in data records

A function to evaluate and log changes in data records between two dataframes and optionally create a markdown changelog.

Usage

changelog(
  x,
  y,
  by = "column",
  report = FALSE,
  file = "CHANGELOG.md",
  author = NULL,
  date = NULL,
  ver = NULL
)

Arguments

x

data.frame. A data.frame to compare against y. Treated as the original version of the data records.

y

data.frame. A data.frame to compare against x. Treated as the replacement version of the data records.

by

character. How should x and y be compared? Options include by "column" (the default) or "row".

report

logical. If TRUE, a verbose report will be written to file.

file

character. Path, name, and extension of the file to be saved. Defaults to current working directory and "CHANGELOG.md".

author

character. The author of to report in the changelog report, if requested.

date

character. The data to report in the changelog report, if requested. By default, the current date is used.

ver

character. The version number to report in the changelog report, if requested.

Value

A data.frame containing data record comparison results. Only changes are reported. If report is TRUE, a changelog report will also be saved to file.

Details

This function assumes that input data.frames maintain the same structure and aims to evaluate whether existing record values have been changed (e.g. through data cleaning).

Examples

# Dataset
x <- mtcars
# Shuffled dataset
y <- x[sample(nrow(x)), ]
# Evaluate differences
changelog(x = x, y = y)
#>     column                 row original replacement
#> 1      mpg           Mazda RX4   21.000      17.300
#> 2      mpg       Mazda RX4 Wag   21.000      15.200
#> 3      mpg          Datsun 710   22.800      16.400
#> 4      mpg      Hornet 4 Drive   21.400      18.700
#> 5      mpg   Hornet Sportabout   18.700      18.100
#> 6      mpg             Valiant   18.100      10.400
#> 7      mpg          Duster 360   14.300      26.000
#> 8      mpg           Merc 240D   24.400      21.400
#> 9      mpg            Merc 230   22.800      30.400
#> 10     mpg            Merc 280   19.200      22.800
#> 11     mpg           Merc 280C   17.800      15.800
#> 12     mpg          Merc 450SE   16.400      15.500
#> 13     mpg          Merc 450SL   17.300      27.300
#> 14     mpg         Merc 450SLC   15.200      21.400
#> 15     mpg  Cadillac Fleetwood   10.400      21.000
#> 16     mpg Lincoln Continental   10.400      17.800
#> 17     mpg   Chrysler Imperial   14.700      15.200
#> 18     mpg            Fiat 128   32.400      14.700
#> 19     mpg         Honda Civic   30.400      21.500
#> 20     mpg      Toyota Corolla   33.900      10.400
#> 21     mpg       Toyota Corona   21.500      19.700
#> 22     mpg    Dodge Challenger   15.500      32.400
#> 23     mpg         AMC Javelin   15.200      13.300
#> 24     mpg          Camaro Z28   13.300      19.200
#> 25     mpg    Pontiac Firebird   19.200      24.400
#> 26     mpg           Fiat X1-9   27.300      33.900
#> 27     mpg       Porsche 914-2   26.000      22.800
#> 28     mpg        Lotus Europa   30.400      21.000
#> 29     mpg      Ford Pantera L   15.800      14.300
#> 30     mpg        Ferrari Dino   19.700      19.200
#> 31     mpg       Maserati Bora   15.000      30.400
#> 32     mpg          Volvo 142E   21.400      15.000
#> 33     cyl           Mazda RX4    6.000       8.000
#> 34     cyl       Mazda RX4 Wag    6.000       8.000
#> 35     cyl          Datsun 710    4.000       8.000
#> 36     cyl      Hornet 4 Drive    6.000       8.000
#> 37     cyl   Hornet Sportabout    8.000       6.000
#> 38     cyl             Valiant    6.000       8.000
#> 39     cyl          Duster 360    8.000       4.000
#> 40     cyl           Merc 240D    4.000       6.000
#> 41     cyl            Merc 280    6.000       4.000
#> 42     cyl           Merc 280C    6.000       8.000
#> 43     cyl          Merc 450SL    8.000       4.000
#> 44     cyl         Merc 450SLC    8.000       4.000
#> 45     cyl  Cadillac Fleetwood    8.000       6.000
#> 46     cyl Lincoln Continental    8.000       6.000
#> 47     cyl            Fiat 128    4.000       8.000
#> 48     cyl      Toyota Corolla    4.000       8.000
#> 49     cyl       Toyota Corona    4.000       6.000
#> 50     cyl    Dodge Challenger    8.000       4.000
#> 51     cyl    Pontiac Firebird    8.000       4.000
#> 52     cyl        Lotus Europa    4.000       6.000
#> 53     cyl       Maserati Bora    8.000       4.000
#> 54     cyl          Volvo 142E    4.000       8.000
#> 55    disp           Mazda RX4  160.000     275.800
#> 56    disp       Mazda RX4 Wag  160.000     304.000
#> 57    disp          Datsun 710  108.000     275.800
#> 58    disp      Hornet 4 Drive  258.000     360.000
#> 59    disp   Hornet Sportabout  360.000     225.000
#> 60    disp             Valiant  225.000     472.000
#> 61    disp          Duster 360  360.000     120.300
#> 62    disp           Merc 240D  146.700     258.000
#> 63    disp            Merc 230  140.800      95.100
#> 64    disp            Merc 280  167.600     140.800
#> 65    disp           Merc 280C  167.600     351.000
#> 66    disp          Merc 450SE  275.800     318.000
#> 67    disp          Merc 450SL  275.800      79.000
#> 68    disp         Merc 450SLC  275.800     121.000
#> 69    disp  Cadillac Fleetwood  472.000     160.000
#> 70    disp Lincoln Continental  460.000     167.600
#> 71    disp   Chrysler Imperial  440.000     275.800
#> 72    disp            Fiat 128   78.700     440.000
#> 73    disp         Honda Civic   75.700     120.100
#> 74    disp      Toyota Corolla   71.100     460.000
#> 75    disp       Toyota Corona  120.100     145.000
#> 76    disp    Dodge Challenger  318.000      78.700
#> 77    disp         AMC Javelin  304.000     350.000
#> 78    disp          Camaro Z28  350.000     400.000
#> 79    disp    Pontiac Firebird  400.000     146.700
#> 80    disp           Fiat X1-9   79.000      71.100
#> 81    disp       Porsche 914-2  120.300     108.000
#> 82    disp        Lotus Europa   95.100     160.000
#> 83    disp      Ford Pantera L  351.000     360.000
#> 84    disp        Ferrari Dino  145.000     167.600
#> 85    disp       Maserati Bora  301.000      75.700
#> 86    disp          Volvo 142E  121.000     301.000
#> 87      hp           Mazda RX4  110.000     180.000
#> 88      hp       Mazda RX4 Wag  110.000     150.000
#> 89      hp          Datsun 710   93.000     180.000
#> 90      hp      Hornet 4 Drive  110.000     175.000
#> 91      hp   Hornet Sportabout  175.000     105.000
#> 92      hp             Valiant  105.000     205.000
#> 93      hp          Duster 360  245.000      91.000
#> 94      hp           Merc 240D   62.000     110.000
#> 95      hp            Merc 230   95.000     113.000
#> 96      hp            Merc 280  123.000      95.000
#> 97      hp           Merc 280C  123.000     264.000
#> 98      hp          Merc 450SE  180.000     150.000
#> 99      hp          Merc 450SL  180.000      66.000
#> 100     hp         Merc 450SLC  180.000     109.000
#> 101     hp  Cadillac Fleetwood  205.000     110.000
#> 102     hp Lincoln Continental  215.000     123.000
#> 103     hp   Chrysler Imperial  230.000     180.000
#> 104     hp            Fiat 128   66.000     230.000
#> 105     hp         Honda Civic   52.000      97.000
#> 106     hp      Toyota Corolla   65.000     215.000
#> 107     hp       Toyota Corona   97.000     175.000
#> 108     hp    Dodge Challenger  150.000      66.000
#> 109     hp         AMC Javelin  150.000     245.000
#> 110     hp          Camaro Z28  245.000     175.000
#> 111     hp    Pontiac Firebird  175.000      62.000
#> 112     hp           Fiat X1-9   66.000      65.000
#> 113     hp       Porsche 914-2   91.000      93.000
#> 114     hp        Lotus Europa  113.000     110.000
#> 115     hp      Ford Pantera L  264.000     245.000
#> 116     hp        Ferrari Dino  175.000     123.000
#> 117     hp       Maserati Bora  335.000      52.000
#> 118     hp          Volvo 142E  109.000     335.000
#> 119   drat           Mazda RX4    3.900       3.070
#> 120   drat       Mazda RX4 Wag    3.900       3.150
#> 121   drat          Datsun 710    3.850       3.070
#> 122   drat      Hornet 4 Drive    3.080       3.150
#> 123   drat   Hornet Sportabout    3.150       2.760
#> 124   drat             Valiant    2.760       2.930
#> 125   drat          Duster 360    3.210       4.430
#> 126   drat           Merc 240D    3.690       3.080
#> 127   drat            Merc 230    3.920       3.770
#> 128   drat           Merc 280C    3.920       4.220
#> 129   drat          Merc 450SE    3.070       2.760
#> 130   drat          Merc 450SL    3.070       4.080
#> 131   drat         Merc 450SLC    3.070       4.110
#> 132   drat  Cadillac Fleetwood    2.930       3.900
#> 133   drat Lincoln Continental    3.000       3.920
#> 134   drat   Chrysler Imperial    3.230       3.070
#> 135   drat            Fiat 128    4.080       3.230
#> 136   drat         Honda Civic    4.930       3.700
#> 137   drat      Toyota Corolla    4.220       3.000
#> 138   drat       Toyota Corona    3.700       3.620
#> 139   drat    Dodge Challenger    2.760       4.080
#> 140   drat         AMC Javelin    3.150       3.730
#> 141   drat          Camaro Z28    3.730       3.080
#> 142   drat    Pontiac Firebird    3.080       3.690
#> 143   drat           Fiat X1-9    4.080       4.220
#> 144   drat       Porsche 914-2    4.430       3.850
#> 145   drat        Lotus Europa    3.770       3.900
#> 146   drat      Ford Pantera L    4.220       3.210
#> 147   drat        Ferrari Dino    3.620       3.920
#> 148   drat       Maserati Bora    3.540       4.930
#> 149   drat          Volvo 142E    4.110       3.540
#> 150     wt           Mazda RX4    2.620       3.730
#> 151     wt       Mazda RX4 Wag    2.875       3.435
#> 152     wt          Datsun 710    2.320       4.070
#> 153     wt      Hornet 4 Drive    3.215       3.440
#> 154     wt   Hornet Sportabout    3.440       3.460
#> 155     wt             Valiant    3.460       5.250
#> 156     wt          Duster 360    3.570       2.140
#> 157     wt           Merc 240D    3.190       3.215
#> 158     wt            Merc 230    3.150       1.513
#> 159     wt            Merc 280    3.440       3.150
#> 160     wt           Merc 280C    3.440       3.170
#> 161     wt          Merc 450SE    4.070       3.520
#> 162     wt          Merc 450SL    3.730       1.935
#> 163     wt         Merc 450SLC    3.780       2.780
#> 164     wt  Cadillac Fleetwood    5.250       2.875
#> 165     wt Lincoln Continental    5.424       3.440
#> 166     wt   Chrysler Imperial    5.345       3.780
#> 167     wt            Fiat 128    2.200       5.345
#> 168     wt         Honda Civic    1.615       2.465
#> 169     wt      Toyota Corolla    1.835       5.424
#> 170     wt       Toyota Corona    2.465       2.770
#> 171     wt    Dodge Challenger    3.520       2.200
#> 172     wt         AMC Javelin    3.435       3.840
#> 173     wt          Camaro Z28    3.840       3.845
#> 174     wt    Pontiac Firebird    3.845       3.190
#> 175     wt           Fiat X1-9    1.935       1.835
#> 176     wt       Porsche 914-2    2.140       2.320
#> 177     wt        Lotus Europa    1.513       2.620
#> 178     wt      Ford Pantera L    3.170       3.570
#> 179     wt        Ferrari Dino    2.770       3.440
#> 180     wt       Maserati Bora    3.570       1.615
#> 181     wt          Volvo 142E    2.780       3.570
#> 182   qsec           Mazda RX4   16.460      17.600
#> 183   qsec       Mazda RX4 Wag   17.020      17.300
#> 184   qsec          Datsun 710   18.610      17.400
#> 185   qsec      Hornet 4 Drive   19.440      17.020
#> 186   qsec   Hornet Sportabout   17.020      20.220
#> 187   qsec             Valiant   20.220      17.980
#> 188   qsec          Duster 360   15.840      16.700
#> 189   qsec           Merc 240D   20.000      19.440
#> 190   qsec            Merc 230   22.900      16.900
#> 191   qsec            Merc 280   18.300      22.900
#> 192   qsec           Merc 280C   18.900      14.500
#> 193   qsec          Merc 450SE   17.400      16.870
#> 194   qsec          Merc 450SL   17.600      18.900
#> 195   qsec         Merc 450SLC   18.000      18.600
#> 196   qsec  Cadillac Fleetwood   17.980      17.020
#> 197   qsec Lincoln Continental   17.820      18.900
#> 198   qsec   Chrysler Imperial   17.420      18.000
#> 199   qsec            Fiat 128   19.470      17.420
#> 200   qsec         Honda Civic   18.520      20.010
#> 201   qsec      Toyota Corolla   19.900      17.820
#> 202   qsec       Toyota Corona   20.010      15.500
#> 203   qsec    Dodge Challenger   16.870      19.470
#> 204   qsec         AMC Javelin   17.300      15.410
#> 205   qsec          Camaro Z28   15.410      17.050
#> 206   qsec    Pontiac Firebird   17.050      20.000
#> 207   qsec           Fiat X1-9   18.900      19.900
#> 208   qsec       Porsche 914-2   16.700      18.610
#> 209   qsec        Lotus Europa   16.900      16.460
#> 210   qsec      Ford Pantera L   14.500      15.840
#> 211   qsec        Ferrari Dino   15.500      18.300
#> 212   qsec       Maserati Bora   14.600      18.520
#> 213   qsec          Volvo 142E   18.600      14.600
#> 214     vs          Datsun 710    1.000       0.000
#> 215     vs      Hornet 4 Drive    1.000       0.000
#> 216     vs   Hornet Sportabout    0.000       1.000
#> 217     vs             Valiant    1.000       0.000
#> 218     vs           Merc 280C    1.000       0.000
#> 219     vs          Merc 450SL    0.000       1.000
#> 220     vs         Merc 450SLC    0.000       1.000
#> 221     vs Lincoln Continental    0.000       1.000
#> 222     vs            Fiat 128    1.000       0.000
#> 223     vs      Toyota Corolla    1.000       0.000
#> 224     vs       Toyota Corona    1.000       0.000
#> 225     vs    Dodge Challenger    0.000       1.000
#> 226     vs    Pontiac Firebird    0.000       1.000
#> 227     vs       Porsche 914-2    0.000       1.000
#> 228     vs        Lotus Europa    1.000       0.000
#> 229     vs        Ferrari Dino    0.000       1.000
#> 230     vs       Maserati Bora    0.000       1.000
#> 231     vs          Volvo 142E    1.000       0.000
#> 232     am           Mazda RX4    1.000       0.000
#> 233     am       Mazda RX4 Wag    1.000       0.000
#> 234     am          Datsun 710    1.000       0.000
#> 235     am          Duster 360    0.000       1.000
#> 236     am            Merc 230    0.000       1.000
#> 237     am           Merc 280C    0.000       1.000
#> 238     am          Merc 450SL    0.000       1.000
#> 239     am         Merc 450SLC    0.000       1.000
#> 240     am  Cadillac Fleetwood    0.000       1.000
#> 241     am            Fiat 128    1.000       0.000
#> 242     am         Honda Civic    1.000       0.000
#> 243     am      Toyota Corolla    1.000       0.000
#> 244     am       Toyota Corona    0.000       1.000
#> 245     am    Dodge Challenger    0.000       1.000
#> 246     am      Ford Pantera L    1.000       0.000
#> 247     am        Ferrari Dino    1.000       0.000
#> 248   gear           Mazda RX4    4.000       3.000
#> 249   gear       Mazda RX4 Wag    4.000       3.000
#> 250   gear          Datsun 710    4.000       3.000
#> 251   gear          Duster 360    3.000       5.000
#> 252   gear           Merc 240D    4.000       3.000
#> 253   gear            Merc 230    4.000       5.000
#> 254   gear           Merc 280C    4.000       5.000
#> 255   gear          Merc 450SL    3.000       4.000
#> 256   gear         Merc 450SLC    3.000       4.000
#> 257   gear  Cadillac Fleetwood    3.000       4.000
#> 258   gear Lincoln Continental    3.000       4.000
#> 259   gear            Fiat 128    4.000       3.000
#> 260   gear         Honda Civic    4.000       3.000
#> 261   gear      Toyota Corolla    4.000       3.000
#> 262   gear       Toyota Corona    3.000       5.000
#> 263   gear    Dodge Challenger    3.000       4.000
#> 264   gear    Pontiac Firebird    3.000       4.000
#> 265   gear       Porsche 914-2    5.000       4.000
#> 266   gear        Lotus Europa    5.000       4.000
#> 267   gear      Ford Pantera L    5.000       3.000
#> 268   gear        Ferrari Dino    5.000       4.000
#> 269   gear       Maserati Bora    5.000       4.000
#> 270   gear          Volvo 142E    4.000       5.000
#> 271   carb           Mazda RX4    4.000       3.000
#> 272   carb       Mazda RX4 Wag    4.000       2.000
#> 273   carb          Datsun 710    1.000       3.000
#> 274   carb      Hornet 4 Drive    1.000       2.000
#> 275   carb   Hornet Sportabout    2.000       1.000
#> 276   carb             Valiant    1.000       4.000
#> 277   carb          Duster 360    4.000       2.000
#> 278   carb           Merc 240D    2.000       1.000
#> 279   carb            Merc 280    4.000       2.000
#> 280   carb          Merc 450SE    3.000       2.000
#> 281   carb          Merc 450SL    3.000       1.000
#> 282   carb         Merc 450SLC    3.000       2.000
#> 283   carb   Chrysler Imperial    4.000       3.000
#> 284   carb            Fiat 128    1.000       4.000
#> 285   carb         Honda Civic    2.000       1.000
#> 286   carb      Toyota Corolla    1.000       4.000
#> 287   carb       Toyota Corona    1.000       6.000
#> 288   carb    Dodge Challenger    2.000       1.000
#> 289   carb         AMC Javelin    2.000       4.000
#> 290   carb          Camaro Z28    4.000       2.000
#> 291   carb       Porsche 914-2    2.000       1.000
#> 292   carb        Lotus Europa    2.000       4.000
#> 293   carb        Ferrari Dino    6.000       4.000
#> 294   carb       Maserati Bora    8.000       2.000
#> 295   carb          Volvo 142E    2.000       8.000
# Create markdown report
changelog(x = x, y = y, report = TRUE, author = "Bob", ver = "0.0.1")
#>     column                 row original replacement
#> 1      mpg           Mazda RX4   21.000      17.300
#> 2      mpg       Mazda RX4 Wag   21.000      15.200
#> 3      mpg          Datsun 710   22.800      16.400
#> 4      mpg      Hornet 4 Drive   21.400      18.700
#> 5      mpg   Hornet Sportabout   18.700      18.100
#> 6      mpg             Valiant   18.100      10.400
#> 7      mpg          Duster 360   14.300      26.000
#> 8      mpg           Merc 240D   24.400      21.400
#> 9      mpg            Merc 230   22.800      30.400
#> 10     mpg            Merc 280   19.200      22.800
#> 11     mpg           Merc 280C   17.800      15.800
#> 12     mpg          Merc 450SE   16.400      15.500
#> 13     mpg          Merc 450SL   17.300      27.300
#> 14     mpg         Merc 450SLC   15.200      21.400
#> 15     mpg  Cadillac Fleetwood   10.400      21.000
#> 16     mpg Lincoln Continental   10.400      17.800
#> 17     mpg   Chrysler Imperial   14.700      15.200
#> 18     mpg            Fiat 128   32.400      14.700
#> 19     mpg         Honda Civic   30.400      21.500
#> 20     mpg      Toyota Corolla   33.900      10.400
#> 21     mpg       Toyota Corona   21.500      19.700
#> 22     mpg    Dodge Challenger   15.500      32.400
#> 23     mpg         AMC Javelin   15.200      13.300
#> 24     mpg          Camaro Z28   13.300      19.200
#> 25     mpg    Pontiac Firebird   19.200      24.400
#> 26     mpg           Fiat X1-9   27.300      33.900
#> 27     mpg       Porsche 914-2   26.000      22.800
#> 28     mpg        Lotus Europa   30.400      21.000
#> 29     mpg      Ford Pantera L   15.800      14.300
#> 30     mpg        Ferrari Dino   19.700      19.200
#> 31     mpg       Maserati Bora   15.000      30.400
#> 32     mpg          Volvo 142E   21.400      15.000
#> 33     cyl           Mazda RX4    6.000       8.000
#> 34     cyl       Mazda RX4 Wag    6.000       8.000
#> 35     cyl          Datsun 710    4.000       8.000
#> 36     cyl      Hornet 4 Drive    6.000       8.000
#> 37     cyl   Hornet Sportabout    8.000       6.000
#> 38     cyl             Valiant    6.000       8.000
#> 39     cyl          Duster 360    8.000       4.000
#> 40     cyl           Merc 240D    4.000       6.000
#> 41     cyl            Merc 280    6.000       4.000
#> 42     cyl           Merc 280C    6.000       8.000
#> 43     cyl          Merc 450SL    8.000       4.000
#> 44     cyl         Merc 450SLC    8.000       4.000
#> 45     cyl  Cadillac Fleetwood    8.000       6.000
#> 46     cyl Lincoln Continental    8.000       6.000
#> 47     cyl            Fiat 128    4.000       8.000
#> 48     cyl      Toyota Corolla    4.000       8.000
#> 49     cyl       Toyota Corona    4.000       6.000
#> 50     cyl    Dodge Challenger    8.000       4.000
#> 51     cyl    Pontiac Firebird    8.000       4.000
#> 52     cyl        Lotus Europa    4.000       6.000
#> 53     cyl       Maserati Bora    8.000       4.000
#> 54     cyl          Volvo 142E    4.000       8.000
#> 55    disp           Mazda RX4  160.000     275.800
#> 56    disp       Mazda RX4 Wag  160.000     304.000
#> 57    disp          Datsun 710  108.000     275.800
#> 58    disp      Hornet 4 Drive  258.000     360.000
#> 59    disp   Hornet Sportabout  360.000     225.000
#> 60    disp             Valiant  225.000     472.000
#> 61    disp          Duster 360  360.000     120.300
#> 62    disp           Merc 240D  146.700     258.000
#> 63    disp            Merc 230  140.800      95.100
#> 64    disp            Merc 280  167.600     140.800
#> 65    disp           Merc 280C  167.600     351.000
#> 66    disp          Merc 450SE  275.800     318.000
#> 67    disp          Merc 450SL  275.800      79.000
#> 68    disp         Merc 450SLC  275.800     121.000
#> 69    disp  Cadillac Fleetwood  472.000     160.000
#> 70    disp Lincoln Continental  460.000     167.600
#> 71    disp   Chrysler Imperial  440.000     275.800
#> 72    disp            Fiat 128   78.700     440.000
#> 73    disp         Honda Civic   75.700     120.100
#> 74    disp      Toyota Corolla   71.100     460.000
#> 75    disp       Toyota Corona  120.100     145.000
#> 76    disp    Dodge Challenger  318.000      78.700
#> 77    disp         AMC Javelin  304.000     350.000
#> 78    disp          Camaro Z28  350.000     400.000
#> 79    disp    Pontiac Firebird  400.000     146.700
#> 80    disp           Fiat X1-9   79.000      71.100
#> 81    disp       Porsche 914-2  120.300     108.000
#> 82    disp        Lotus Europa   95.100     160.000
#> 83    disp      Ford Pantera L  351.000     360.000
#> 84    disp        Ferrari Dino  145.000     167.600
#> 85    disp       Maserati Bora  301.000      75.700
#> 86    disp          Volvo 142E  121.000     301.000
#> 87      hp           Mazda RX4  110.000     180.000
#> 88      hp       Mazda RX4 Wag  110.000     150.000
#> 89      hp          Datsun 710   93.000     180.000
#> 90      hp      Hornet 4 Drive  110.000     175.000
#> 91      hp   Hornet Sportabout  175.000     105.000
#> 92      hp             Valiant  105.000     205.000
#> 93      hp          Duster 360  245.000      91.000
#> 94      hp           Merc 240D   62.000     110.000
#> 95      hp            Merc 230   95.000     113.000
#> 96      hp            Merc 280  123.000      95.000
#> 97      hp           Merc 280C  123.000     264.000
#> 98      hp          Merc 450SE  180.000     150.000
#> 99      hp          Merc 450SL  180.000      66.000
#> 100     hp         Merc 450SLC  180.000     109.000
#> 101     hp  Cadillac Fleetwood  205.000     110.000
#> 102     hp Lincoln Continental  215.000     123.000
#> 103     hp   Chrysler Imperial  230.000     180.000
#> 104     hp            Fiat 128   66.000     230.000
#> 105     hp         Honda Civic   52.000      97.000
#> 106     hp      Toyota Corolla   65.000     215.000
#> 107     hp       Toyota Corona   97.000     175.000
#> 108     hp    Dodge Challenger  150.000      66.000
#> 109     hp         AMC Javelin  150.000     245.000
#> 110     hp          Camaro Z28  245.000     175.000
#> 111     hp    Pontiac Firebird  175.000      62.000
#> 112     hp           Fiat X1-9   66.000      65.000
#> 113     hp       Porsche 914-2   91.000      93.000
#> 114     hp        Lotus Europa  113.000     110.000
#> 115     hp      Ford Pantera L  264.000     245.000
#> 116     hp        Ferrari Dino  175.000     123.000
#> 117     hp       Maserati Bora  335.000      52.000
#> 118     hp          Volvo 142E  109.000     335.000
#> 119   drat           Mazda RX4    3.900       3.070
#> 120   drat       Mazda RX4 Wag    3.900       3.150
#> 121   drat          Datsun 710    3.850       3.070
#> 122   drat      Hornet 4 Drive    3.080       3.150
#> 123   drat   Hornet Sportabout    3.150       2.760
#> 124   drat             Valiant    2.760       2.930
#> 125   drat          Duster 360    3.210       4.430
#> 126   drat           Merc 240D    3.690       3.080
#> 127   drat            Merc 230    3.920       3.770
#> 128   drat           Merc 280C    3.920       4.220
#> 129   drat          Merc 450SE    3.070       2.760
#> 130   drat          Merc 450SL    3.070       4.080
#> 131   drat         Merc 450SLC    3.070       4.110
#> 132   drat  Cadillac Fleetwood    2.930       3.900
#> 133   drat Lincoln Continental    3.000       3.920
#> 134   drat   Chrysler Imperial    3.230       3.070
#> 135   drat            Fiat 128    4.080       3.230
#> 136   drat         Honda Civic    4.930       3.700
#> 137   drat      Toyota Corolla    4.220       3.000
#> 138   drat       Toyota Corona    3.700       3.620
#> 139   drat    Dodge Challenger    2.760       4.080
#> 140   drat         AMC Javelin    3.150       3.730
#> 141   drat          Camaro Z28    3.730       3.080
#> 142   drat    Pontiac Firebird    3.080       3.690
#> 143   drat           Fiat X1-9    4.080       4.220
#> 144   drat       Porsche 914-2    4.430       3.850
#> 145   drat        Lotus Europa    3.770       3.900
#> 146   drat      Ford Pantera L    4.220       3.210
#> 147   drat        Ferrari Dino    3.620       3.920
#> 148   drat       Maserati Bora    3.540       4.930
#> 149   drat          Volvo 142E    4.110       3.540
#> 150     wt           Mazda RX4    2.620       3.730
#> 151     wt       Mazda RX4 Wag    2.875       3.435
#> 152     wt          Datsun 710    2.320       4.070
#> 153     wt      Hornet 4 Drive    3.215       3.440
#> 154     wt   Hornet Sportabout    3.440       3.460
#> 155     wt             Valiant    3.460       5.250
#> 156     wt          Duster 360    3.570       2.140
#> 157     wt           Merc 240D    3.190       3.215
#> 158     wt            Merc 230    3.150       1.513
#> 159     wt            Merc 280    3.440       3.150
#> 160     wt           Merc 280C    3.440       3.170
#> 161     wt          Merc 450SE    4.070       3.520
#> 162     wt          Merc 450SL    3.730       1.935
#> 163     wt         Merc 450SLC    3.780       2.780
#> 164     wt  Cadillac Fleetwood    5.250       2.875
#> 165     wt Lincoln Continental    5.424       3.440
#> 166     wt   Chrysler Imperial    5.345       3.780
#> 167     wt            Fiat 128    2.200       5.345
#> 168     wt         Honda Civic    1.615       2.465
#> 169     wt      Toyota Corolla    1.835       5.424
#> 170     wt       Toyota Corona    2.465       2.770
#> 171     wt    Dodge Challenger    3.520       2.200
#> 172     wt         AMC Javelin    3.435       3.840
#> 173     wt          Camaro Z28    3.840       3.845
#> 174     wt    Pontiac Firebird    3.845       3.190
#> 175     wt           Fiat X1-9    1.935       1.835
#> 176     wt       Porsche 914-2    2.140       2.320
#> 177     wt        Lotus Europa    1.513       2.620
#> 178     wt      Ford Pantera L    3.170       3.570
#> 179     wt        Ferrari Dino    2.770       3.440
#> 180     wt       Maserati Bora    3.570       1.615
#> 181     wt          Volvo 142E    2.780       3.570
#> 182   qsec           Mazda RX4   16.460      17.600
#> 183   qsec       Mazda RX4 Wag   17.020      17.300
#> 184   qsec          Datsun 710   18.610      17.400
#> 185   qsec      Hornet 4 Drive   19.440      17.020
#> 186   qsec   Hornet Sportabout   17.020      20.220
#> 187   qsec             Valiant   20.220      17.980
#> 188   qsec          Duster 360   15.840      16.700
#> 189   qsec           Merc 240D   20.000      19.440
#> 190   qsec            Merc 230   22.900      16.900
#> 191   qsec            Merc 280   18.300      22.900
#> 192   qsec           Merc 280C   18.900      14.500
#> 193   qsec          Merc 450SE   17.400      16.870
#> 194   qsec          Merc 450SL   17.600      18.900
#> 195   qsec         Merc 450SLC   18.000      18.600
#> 196   qsec  Cadillac Fleetwood   17.980      17.020
#> 197   qsec Lincoln Continental   17.820      18.900
#> 198   qsec   Chrysler Imperial   17.420      18.000
#> 199   qsec            Fiat 128   19.470      17.420
#> 200   qsec         Honda Civic   18.520      20.010
#> 201   qsec      Toyota Corolla   19.900      17.820
#> 202   qsec       Toyota Corona   20.010      15.500
#> 203   qsec    Dodge Challenger   16.870      19.470
#> 204   qsec         AMC Javelin   17.300      15.410
#> 205   qsec          Camaro Z28   15.410      17.050
#> 206   qsec    Pontiac Firebird   17.050      20.000
#> 207   qsec           Fiat X1-9   18.900      19.900
#> 208   qsec       Porsche 914-2   16.700      18.610
#> 209   qsec        Lotus Europa   16.900      16.460
#> 210   qsec      Ford Pantera L   14.500      15.840
#> 211   qsec        Ferrari Dino   15.500      18.300
#> 212   qsec       Maserati Bora   14.600      18.520
#> 213   qsec          Volvo 142E   18.600      14.600
#> 214     vs          Datsun 710    1.000       0.000
#> 215     vs      Hornet 4 Drive    1.000       0.000
#> 216     vs   Hornet Sportabout    0.000       1.000
#> 217     vs             Valiant    1.000       0.000
#> 218     vs           Merc 280C    1.000       0.000
#> 219     vs          Merc 450SL    0.000       1.000
#> 220     vs         Merc 450SLC    0.000       1.000
#> 221     vs Lincoln Continental    0.000       1.000
#> 222     vs            Fiat 128    1.000       0.000
#> 223     vs      Toyota Corolla    1.000       0.000
#> 224     vs       Toyota Corona    1.000       0.000
#> 225     vs    Dodge Challenger    0.000       1.000
#> 226     vs    Pontiac Firebird    0.000       1.000
#> 227     vs       Porsche 914-2    0.000       1.000
#> 228     vs        Lotus Europa    1.000       0.000
#> 229     vs        Ferrari Dino    0.000       1.000
#> 230     vs       Maserati Bora    0.000       1.000
#> 231     vs          Volvo 142E    1.000       0.000
#> 232     am           Mazda RX4    1.000       0.000
#> 233     am       Mazda RX4 Wag    1.000       0.000
#> 234     am          Datsun 710    1.000       0.000
#> 235     am          Duster 360    0.000       1.000
#> 236     am            Merc 230    0.000       1.000
#> 237     am           Merc 280C    0.000       1.000
#> 238     am          Merc 450SL    0.000       1.000
#> 239     am         Merc 450SLC    0.000       1.000
#> 240     am  Cadillac Fleetwood    0.000       1.000
#> 241     am            Fiat 128    1.000       0.000
#> 242     am         Honda Civic    1.000       0.000
#> 243     am      Toyota Corolla    1.000       0.000
#> 244     am       Toyota Corona    0.000       1.000
#> 245     am    Dodge Challenger    0.000       1.000
#> 246     am      Ford Pantera L    1.000       0.000
#> 247     am        Ferrari Dino    1.000       0.000
#> 248   gear           Mazda RX4    4.000       3.000
#> 249   gear       Mazda RX4 Wag    4.000       3.000
#> 250   gear          Datsun 710    4.000       3.000
#> 251   gear          Duster 360    3.000       5.000
#> 252   gear           Merc 240D    4.000       3.000
#> 253   gear            Merc 230    4.000       5.000
#> 254   gear           Merc 280C    4.000       5.000
#> 255   gear          Merc 450SL    3.000       4.000
#> 256   gear         Merc 450SLC    3.000       4.000
#> 257   gear  Cadillac Fleetwood    3.000       4.000
#> 258   gear Lincoln Continental    3.000       4.000
#> 259   gear            Fiat 128    4.000       3.000
#> 260   gear         Honda Civic    4.000       3.000
#> 261   gear      Toyota Corolla    4.000       3.000
#> 262   gear       Toyota Corona    3.000       5.000
#> 263   gear    Dodge Challenger    3.000       4.000
#> 264   gear    Pontiac Firebird    3.000       4.000
#> 265   gear       Porsche 914-2    5.000       4.000
#> 266   gear        Lotus Europa    5.000       4.000
#> 267   gear      Ford Pantera L    5.000       3.000
#> 268   gear        Ferrari Dino    5.000       4.000
#> 269   gear       Maserati Bora    5.000       4.000
#> 270   gear          Volvo 142E    4.000       5.000
#> 271   carb           Mazda RX4    4.000       3.000
#> 272   carb       Mazda RX4 Wag    4.000       2.000
#> 273   carb          Datsun 710    1.000       3.000
#> 274   carb      Hornet 4 Drive    1.000       2.000
#> 275   carb   Hornet Sportabout    2.000       1.000
#> 276   carb             Valiant    1.000       4.000
#> 277   carb          Duster 360    4.000       2.000
#> 278   carb           Merc 240D    2.000       1.000
#> 279   carb            Merc 280    4.000       2.000
#> 280   carb          Merc 450SE    3.000       2.000
#> 281   carb          Merc 450SL    3.000       1.000
#> 282   carb         Merc 450SLC    3.000       2.000
#> 283   carb   Chrysler Imperial    4.000       3.000
#> 284   carb            Fiat 128    1.000       4.000
#> 285   carb         Honda Civic    2.000       1.000
#> 286   carb      Toyota Corolla    1.000       4.000
#> 287   carb       Toyota Corona    1.000       6.000
#> 288   carb    Dodge Challenger    2.000       1.000
#> 289   carb         AMC Javelin    2.000       4.000
#> 290   carb          Camaro Z28    4.000       2.000
#> 291   carb       Porsche 914-2    2.000       1.000
#> 292   carb        Lotus Europa    2.000       4.000
#> 293   carb        Ferrari Dino    6.000       4.000
#> 294   carb       Maserati Bora    8.000       2.000
#> 295   carb          Volvo 142E    2.000       8.000