Evaluate and log changes in data records

A function to evaluate and log changes in data records between two data.frame and optionally create a markdown changelog report.

Usage

evaluate(
  x,
  y,
  by = "column",
  id = NULL,
  report = FALSE,
  file = "CHANGELOG.md",
  author = NULL,
  date = NULL,
  ver = NULL
)

Arguments

x

data.frame. A data.frame to compare against y. Treated as the original version of the data records.

y

data.frame. A data.frame to compare against x. Treated as the replacement version of the data records.

by

character. How should x and y be compared? Options include by "column" (the default), "row", or "id".

id

character. The name of the column to treat as unique identifiers for comparing x and y. Only applicable if by is set to "id".

report

logical. If TRUE, a verbose report will be written to file.

file

character. Path, name, and extension of the file to be saved. Defaults to current working directory and "CHANGELOG.md".

author

character. The author of to report in the changelog report, if requested.

date

character. The data to report in the changelog report, if requested. By default, the current date is used.

ver

character. The version number to report in the changelog report, if requested.

Value

A data.frame containing data record comparison results. Only changes are reported. If report is TRUE, a changelog report will also be saved to file.

Details

This function assumes that input data.frames maintain the same structure and aims to evaluate whether existing record values have been changed (e.g. through data cleaning).

Examples

# Dataset
x <- mtcars
# Shuffled dataset
y <- x[sample(nrow(x)), ]
# Evaluate differences
evaluate(x = x, y = y)
#>     column                 row original replacement
#> 1      mpg           Mazda RX4   21.000      16.400
#> 2      mpg       Mazda RX4 Wag   21.000      15.000
#> 3      mpg          Datsun 710   22.800      18.700
#> 4      mpg      Hornet 4 Drive   21.400      18.100
#> 5      mpg   Hornet Sportabout   18.700      10.400
#> 6      mpg             Valiant   18.100      30.400
#> 7      mpg          Duster 360   14.300      21.400
#> 8      mpg           Merc 240D   24.400      15.800
#> 9      mpg            Merc 280   19.200      19.700
#> 10     mpg           Merc 280C   17.800      15.200
#> 11     mpg          Merc 450SE   16.400      26.000
#> 12     mpg         Merc 450SLC   15.200      21.000
#> 13     mpg  Cadillac Fleetwood   10.400      17.800
#> 14     mpg Lincoln Continental   10.400      32.400
#> 15     mpg   Chrysler Imperial   14.700      15.500
#> 16     mpg            Fiat 128   32.400      10.400
#> 17     mpg         Honda Civic   30.400      21.400
#> 18     mpg       Toyota Corona   21.500      30.400
#> 19     mpg    Dodge Challenger   15.500      19.200
#> 20     mpg         AMC Javelin   15.200      27.300
#> 21     mpg          Camaro Z28   13.300      15.200
#> 22     mpg    Pontiac Firebird   19.200      22.800
#> 23     mpg           Fiat X1-9   27.300      21.000
#> 24     mpg       Porsche 914-2   26.000      14.700
#> 25     mpg        Lotus Europa   30.400      24.400
#> 26     mpg      Ford Pantera L   15.800      21.500
#> 27     mpg        Ferrari Dino   19.700      14.300
#> 28     mpg       Maserati Bora   15.000      13.300
#> 29     mpg          Volvo 142E   21.400      19.200
#> 30     cyl           Mazda RX4    6.000       8.000
#> 31     cyl       Mazda RX4 Wag    6.000       8.000
#> 32     cyl          Datsun 710    4.000       8.000
#> 33     cyl             Valiant    6.000       4.000
#> 34     cyl          Duster 360    8.000       6.000
#> 35     cyl           Merc 240D    4.000       8.000
#> 36     cyl           Merc 280C    6.000       8.000
#> 37     cyl          Merc 450SE    8.000       4.000
#> 38     cyl         Merc 450SLC    8.000       6.000
#> 39     cyl  Cadillac Fleetwood    8.000       6.000
#> 40     cyl Lincoln Continental    8.000       4.000
#> 41     cyl            Fiat 128    4.000       8.000
#> 42     cyl         AMC Javelin    8.000       4.000
#> 43     cyl    Pontiac Firebird    8.000       4.000
#> 44     cyl           Fiat X1-9    4.000       6.000
#> 45     cyl       Porsche 914-2    4.000       8.000
#> 46     cyl      Ford Pantera L    8.000       4.000
#> 47     cyl        Ferrari Dino    6.000       8.000
#> 48     cyl          Volvo 142E    4.000       6.000
#> 49    disp           Mazda RX4  160.000     275.800
#> 50    disp       Mazda RX4 Wag  160.000     301.000
#> 51    disp          Datsun 710  108.000     360.000
#> 52    disp      Hornet 4 Drive  258.000     225.000
#> 53    disp   Hornet Sportabout  360.000     472.000
#> 54    disp             Valiant  225.000      95.100
#> 55    disp          Duster 360  360.000     258.000
#> 56    disp           Merc 240D  146.700     351.000
#> 57    disp            Merc 280  167.600     145.000
#> 58    disp           Merc 280C  167.600     304.000
#> 59    disp          Merc 450SE  275.800     120.300
#> 60    disp         Merc 450SLC  275.800     160.000
#> 61    disp  Cadillac Fleetwood  472.000     167.600
#> 62    disp Lincoln Continental  460.000      78.700
#> 63    disp   Chrysler Imperial  440.000     318.000
#> 64    disp            Fiat 128   78.700     460.000
#> 65    disp         Honda Civic   75.700     121.000
#> 66    disp       Toyota Corona  120.100      75.700
#> 67    disp    Dodge Challenger  318.000     400.000
#> 68    disp         AMC Javelin  304.000      79.000
#> 69    disp          Camaro Z28  350.000     275.800
#> 70    disp    Pontiac Firebird  400.000     108.000
#> 71    disp           Fiat X1-9   79.000     160.000
#> 72    disp       Porsche 914-2  120.300     440.000
#> 73    disp        Lotus Europa   95.100     146.700
#> 74    disp      Ford Pantera L  351.000     120.100
#> 75    disp        Ferrari Dino  145.000     360.000
#> 76    disp       Maserati Bora  301.000     350.000
#> 77    disp          Volvo 142E  121.000     167.600
#> 78      hp           Mazda RX4  110.000     180.000
#> 79      hp       Mazda RX4 Wag  110.000     335.000
#> 80      hp          Datsun 710   93.000     175.000
#> 81      hp      Hornet 4 Drive  110.000     105.000
#> 82      hp   Hornet Sportabout  175.000     205.000
#> 83      hp             Valiant  105.000     113.000
#> 84      hp          Duster 360  245.000     110.000
#> 85      hp           Merc 240D   62.000     264.000
#> 86      hp            Merc 280  123.000     175.000
#> 87      hp           Merc 280C  123.000     150.000
#> 88      hp          Merc 450SE  180.000      91.000
#> 89      hp         Merc 450SLC  180.000     110.000
#> 90      hp  Cadillac Fleetwood  205.000     123.000
#> 91      hp Lincoln Continental  215.000      66.000
#> 92      hp   Chrysler Imperial  230.000     150.000
#> 93      hp            Fiat 128   66.000     215.000
#> 94      hp         Honda Civic   52.000     109.000
#> 95      hp       Toyota Corona   97.000      52.000
#> 96      hp    Dodge Challenger  150.000     175.000
#> 97      hp         AMC Javelin  150.000      66.000
#> 98      hp          Camaro Z28  245.000     180.000
#> 99      hp    Pontiac Firebird  175.000      93.000
#> 100     hp           Fiat X1-9   66.000     110.000
#> 101     hp       Porsche 914-2   91.000     230.000
#> 102     hp        Lotus Europa  113.000      62.000
#> 103     hp      Ford Pantera L  264.000      97.000
#> 104     hp        Ferrari Dino  175.000     245.000
#> 105     hp       Maserati Bora  335.000     245.000
#> 106     hp          Volvo 142E  109.000     123.000
#> 107   drat           Mazda RX4    3.900       3.070
#> 108   drat       Mazda RX4 Wag    3.900       3.540
#> 109   drat          Datsun 710    3.850       3.150
#> 110   drat      Hornet 4 Drive    3.080       2.760
#> 111   drat   Hornet Sportabout    3.150       2.930
#> 112   drat             Valiant    2.760       3.770
#> 113   drat          Duster 360    3.210       3.080
#> 114   drat           Merc 240D    3.690       4.220
#> 115   drat            Merc 280    3.920       3.620
#> 116   drat           Merc 280C    3.920       3.150
#> 117   drat          Merc 450SE    3.070       4.430
#> 118   drat         Merc 450SLC    3.070       3.900
#> 119   drat  Cadillac Fleetwood    2.930       3.920
#> 120   drat Lincoln Continental    3.000       4.080
#> 121   drat   Chrysler Imperial    3.230       2.760
#> 122   drat            Fiat 128    4.080       3.000
#> 123   drat         Honda Civic    4.930       4.110
#> 124   drat       Toyota Corona    3.700       4.930
#> 125   drat    Dodge Challenger    2.760       3.080
#> 126   drat         AMC Javelin    3.150       4.080
#> 127   drat          Camaro Z28    3.730       3.070
#> 128   drat    Pontiac Firebird    3.080       3.850
#> 129   drat           Fiat X1-9    4.080       3.900
#> 130   drat       Porsche 914-2    4.430       3.230
#> 131   drat        Lotus Europa    3.770       3.690
#> 132   drat      Ford Pantera L    4.220       3.700
#> 133   drat        Ferrari Dino    3.620       3.210
#> 134   drat       Maserati Bora    3.540       3.730
#> 135   drat          Volvo 142E    4.110       3.920
#> 136     wt           Mazda RX4    2.620       4.070
#> 137     wt       Mazda RX4 Wag    2.875       3.570
#> 138     wt          Datsun 710    2.320       3.440
#> 139     wt      Hornet 4 Drive    3.215       3.460
#> 140     wt   Hornet Sportabout    3.440       5.250
#> 141     wt             Valiant    3.460       1.513
#> 142     wt          Duster 360    3.570       3.215
#> 143     wt           Merc 240D    3.190       3.170
#> 144     wt            Merc 280    3.440       2.770
#> 145     wt           Merc 280C    3.440       3.435
#> 146     wt          Merc 450SE    4.070       2.140
#> 147     wt         Merc 450SLC    3.780       2.875
#> 148     wt  Cadillac Fleetwood    5.250       3.440
#> 149     wt Lincoln Continental    5.424       2.200
#> 150     wt   Chrysler Imperial    5.345       3.520
#> 151     wt            Fiat 128    2.200       5.424
#> 152     wt         Honda Civic    1.615       2.780
#> 153     wt       Toyota Corona    2.465       1.615
#> 154     wt    Dodge Challenger    3.520       3.845
#> 155     wt         AMC Javelin    3.435       1.935
#> 156     wt          Camaro Z28    3.840       3.780
#> 157     wt    Pontiac Firebird    3.845       2.320
#> 158     wt           Fiat X1-9    1.935       2.620
#> 159     wt       Porsche 914-2    2.140       5.345
#> 160     wt        Lotus Europa    1.513       3.190
#> 161     wt      Ford Pantera L    3.170       2.465
#> 162     wt        Ferrari Dino    2.770       3.570
#> 163     wt       Maserati Bora    3.570       3.840
#> 164     wt          Volvo 142E    2.780       3.440
#> 165   qsec           Mazda RX4   16.460      17.400
#> 166   qsec       Mazda RX4 Wag   17.020      14.600
#> 167   qsec          Datsun 710   18.610      17.020
#> 168   qsec      Hornet 4 Drive   19.440      20.220
#> 169   qsec   Hornet Sportabout   17.020      17.980
#> 170   qsec             Valiant   20.220      16.900
#> 171   qsec          Duster 360   15.840      19.440
#> 172   qsec           Merc 240D   20.000      14.500
#> 173   qsec            Merc 280   18.300      15.500
#> 174   qsec           Merc 280C   18.900      17.300
#> 175   qsec          Merc 450SE   17.400      16.700
#> 176   qsec         Merc 450SLC   18.000      17.020
#> 177   qsec  Cadillac Fleetwood   17.980      18.900
#> 178   qsec Lincoln Continental   17.820      19.470
#> 179   qsec   Chrysler Imperial   17.420      16.870
#> 180   qsec            Fiat 128   19.470      17.820
#> 181   qsec         Honda Civic   18.520      18.600
#> 182   qsec       Toyota Corona   20.010      18.520
#> 183   qsec    Dodge Challenger   16.870      17.050
#> 184   qsec         AMC Javelin   17.300      18.900
#> 185   qsec          Camaro Z28   15.410      18.000
#> 186   qsec    Pontiac Firebird   17.050      18.610
#> 187   qsec           Fiat X1-9   18.900      16.460
#> 188   qsec       Porsche 914-2   16.700      17.420
#> 189   qsec        Lotus Europa   16.900      20.000
#> 190   qsec      Ford Pantera L   14.500      20.010
#> 191   qsec        Ferrari Dino   15.500      15.840
#> 192   qsec       Maserati Bora   14.600      15.410
#> 193   qsec          Volvo 142E   18.600      18.300
#> 194     vs          Datsun 710    1.000       0.000
#> 195     vs          Duster 360    0.000       1.000
#> 196     vs           Merc 240D    1.000       0.000
#> 197     vs            Merc 280    1.000       0.000
#> 198     vs           Merc 280C    1.000       0.000
#> 199     vs  Cadillac Fleetwood    0.000       1.000
#> 200     vs Lincoln Continental    0.000       1.000
#> 201     vs            Fiat 128    1.000       0.000
#> 202     vs         AMC Javelin    0.000       1.000
#> 203     vs    Pontiac Firebird    0.000       1.000
#> 204     vs           Fiat X1-9    1.000       0.000
#> 205     vs      Ford Pantera L    0.000       1.000
#> 206     am           Mazda RX4    1.000       0.000
#> 207     am          Datsun 710    1.000       0.000
#> 208     am             Valiant    0.000       1.000
#> 209     am           Merc 240D    0.000       1.000
#> 210     am            Merc 280    0.000       1.000
#> 211     am          Merc 450SE    0.000       1.000
#> 212     am         Merc 450SLC    0.000       1.000
#> 213     am Lincoln Continental    0.000       1.000
#> 214     am            Fiat 128    1.000       0.000
#> 215     am       Toyota Corona    0.000       1.000
#> 216     am         AMC Javelin    0.000       1.000
#> 217     am    Pontiac Firebird    0.000       1.000
#> 218     am       Porsche 914-2    1.000       0.000
#> 219     am        Lotus Europa    1.000       0.000
#> 220     am      Ford Pantera L    1.000       0.000
#> 221     am        Ferrari Dino    1.000       0.000
#> 222     am       Maserati Bora    1.000       0.000
#> 223     am          Volvo 142E    1.000       0.000
#> 224   gear           Mazda RX4    4.000       3.000
#> 225   gear       Mazda RX4 Wag    4.000       5.000
#> 226   gear          Datsun 710    4.000       3.000
#> 227   gear             Valiant    3.000       5.000
#> 228   gear           Merc 240D    4.000       5.000
#> 229   gear            Merc 280    4.000       5.000
#> 230   gear           Merc 280C    4.000       3.000
#> 231   gear          Merc 450SE    3.000       5.000
#> 232   gear         Merc 450SLC    3.000       4.000
#> 233   gear  Cadillac Fleetwood    3.000       4.000
#> 234   gear Lincoln Continental    3.000       4.000
#> 235   gear            Fiat 128    4.000       3.000
#> 236   gear       Toyota Corona    3.000       4.000
#> 237   gear         AMC Javelin    3.000       4.000
#> 238   gear    Pontiac Firebird    3.000       4.000
#> 239   gear       Porsche 914-2    5.000       3.000
#> 240   gear        Lotus Europa    5.000       4.000
#> 241   gear      Ford Pantera L    5.000       3.000
#> 242   gear        Ferrari Dino    5.000       3.000
#> 243   gear       Maserati Bora    5.000       3.000
#> 244   carb           Mazda RX4    4.000       3.000
#> 245   carb       Mazda RX4 Wag    4.000       8.000
#> 246   carb          Datsun 710    1.000       2.000
#> 247   carb   Hornet Sportabout    2.000       4.000
#> 248   carb             Valiant    1.000       2.000
#> 249   carb          Duster 360    4.000       1.000
#> 250   carb           Merc 240D    2.000       4.000
#> 251   carb            Merc 280    4.000       6.000
#> 252   carb           Merc 280C    4.000       2.000
#> 253   carb          Merc 450SE    3.000       2.000
#> 254   carb         Merc 450SLC    3.000       4.000
#> 255   carb Lincoln Continental    4.000       1.000
#> 256   carb   Chrysler Imperial    4.000       2.000
#> 257   carb            Fiat 128    1.000       4.000
#> 258   carb       Toyota Corona    1.000       2.000
#> 259   carb         AMC Javelin    2.000       1.000
#> 260   carb          Camaro Z28    4.000       3.000
#> 261   carb    Pontiac Firebird    2.000       1.000
#> 262   carb           Fiat X1-9    1.000       4.000
#> 263   carb       Porsche 914-2    2.000       4.000
#> 264   carb      Ford Pantera L    4.000       1.000
#> 265   carb        Ferrari Dino    6.000       4.000
#> 266   carb       Maserati Bora    8.000       4.000
#> 267   carb          Volvo 142E    2.000       4.000
# Create markdown report
evaluate(x = x, y = y, report = TRUE, author = "Bob", ver = "0.0.1")
#> Error in evaluate(x = x, y = y, report = TRUE, author = "Bob", ver = "0.0.1"): object 'x_y' not found