Evaluate and log changes in data records

A function to evaluate and log changes between two datasets and optionally create a markdown changelog report.

Usage

evaluate(x, y, by, report = FALSE, file = "CHANGELOG.md", metadata = NULL)

Arguments

x

data.frame. The original dataset to compare against y.

y

data.frame. The replacement dataset to compare against x.

by

character. The name of the column to treat as unique record identifiers for comparing x and y.

report

logical. If TRUE, a report of the evaluation of the datasets will be written to file.

file

character. Path, name, and extension of the file to be saved. Defaults to current working directory and "CHANGELOG.md".

metadata

list. A named list of metadata values to include in the header of the changelog report (e.g. author, version), if requested.

Value

A data.frame containing evaluation results. Only changes in data records are reported. If report is TRUE, a changelog report will also be saved to file.

Details

This function evaluates whether records and fields have been added or removed from original (x) and replacement (y) dataset, as well as whether any record values have changed.

Examples

# Original data
x <- mtcars
x$car <- row.names(x)
# Replacement dataset
y <- x
# Change value
y$disp[10] <- 150
# Evaluate differences
evaluate(x = x, y = y, by = "car")
#> Warning: Record values have changed between `x` and `y`.
#>         id column original replacement
#> 1 Merc 280   disp    167.6         150
# Create markdown report
evaluate(x = x, y = y, by = "car", report = TRUE,
         metadata = list(title = "A really cool dataset",
                         author = "Bob",
                         date = Sys.Date(),
                         version = 1))
#> Warning: Record values have changed between `x` and `y`.
#>         id column original replacement
#> 1 Merc 280   disp    167.6         150